check if cuda-keyring is installed before download

+++ .devcontainer/configure.sh
@@ -25,17 +25,19 @@ if [ "$CONTAINER_TYPE" = "with-cuda" ]; then
 
     # Add NVIDIA's package repository to apt so that we can download packages
     # Always use the ubuntu2004 repo because the other repos (e.g., debian11) are missing packages
-    NVIDIA_REPO_URL="https://developer.download.nvidia.com/compute/cuda/repos/ubuntu2004/x86_64"
-    KEYRING_PACKAGE="cuda-keyring_1.0-1_all.deb"
-    KEYRING_PACKAGE_URL="$NVIDIA_REPO_URL/$KEYRING_PACKAGE"
-    KEYRING_PACKAGE_PATH="$(mktemp -d)"
-    KEYRING_PACKAGE_FILE="$KEYRING_PACKAGE_PATH/$KEYRING_PACKAGE"
+    if !(dpkg -s cuda-keyring); then
+        NVIDIA_REPO_URL="https://developer.download.nvidia.com/compute/cuda/repos/ubuntu2004/x86_64"
+        KEYRING_PACKAGE="cuda-keyring_1.1-1_all.deb"
+        KEYRING_PACKAGE_URL="$NVIDIA_REPO_URL/$KEYRING_PACKAGE"
+        KEYRING_PACKAGE_PATH="$(mktemp -d)"
+        KEYRING_PACKAGE_FILE="$KEYRING_PACKAGE_PATH/$KEYRING_PACKAGE"
 
-    wget -O "$KEYRING_PACKAGE_FILE" "$KEYRING_PACKAGE_URL"
-    sudo apt-get install -yq "$KEYRING_PACKAGE_FILE"
-    rm -rf "$KEYRING_PACKAGE_PATH"
+        wget -O "$KEYRING_PACKAGE_FILE" "$KEYRING_PACKAGE_URL"
+        sudo apt-get install -yq "$KEYRING_PACKAGE_FILE"
+        rm -rf "$KEYRING_PACKAGE_PATH"
 
-    sudo apt-get update -yq
+        sudo apt-get update -yq
+    fi
 
     sudo apt-get install -yq "cuda-libraries-${CUDA_VERSION/./-}" "libcudnn8=${CUDNN_VERSION}.*-1+cuda${CUDA_VERSION}" "cuda-nvcc-${CUDA_VERSION/./-}" aptitude
